Refactor CommandManager into CommandParser
Moved extract command and extract parameters methods into the new
CommandParser class and this is created in the Duke class. The
parseCommand will handle the extraction of command and parameters,
before executing the command.

Also created a new GreetCommand class to call the greeting.
