+++ src/main/java/duke/command/CommandType.java
@@ -1,5 +1,5 @@
 package duke.command;
 
-public enum CommandEnum {
+public enum CommandType {
     LIST, DONE, TODO, DEADLINE, EVENT, BYE, DELETE
 }
