+++ src/main/java/duke/command/GreetCommand.java
@@ -1,0 +1,23 @@
+package duke.command;
+
+public class GreetCommand {
+    private static final String GREET = "____________________________________________________________\n" +
+            " __\n" +
+            "('')  Hello! I'm MK.II Titan with feelings.\n" +
+            "// \\  What can I do for you?\n" +
+            "|_/|\n" +
+            "____________________________________________________________\n";
+
+
+    private static final String LOGO = " _    _      _\n" +
+            "| |  | |    | |\n" +
+            "| |  | | ___| | ___ ___  _ __ ___   ___\n" +
+            "| |/\\| |/ _ \\ |/ __/ _ \\| '_ ` _ \\ / _ \\\n" +
+            "\\  /\\  /  __/ | (_| (_) | | | | | |  __/\n" +
+            " \\/  \\/ \\___|_|\\___\\___/|_| |_| |_|\\___|";
+
+    public static void printGreeting() {
+        System.out.println(LOGO);
+        System.out.println(GREET);
+    }
+}
